ProxyServer/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
PyKCS11==1.5.13
requests==2.31.0
